refactor: extract error handling utilities and consolidate agent tools
- Extract common error handling logic to internal/error-handling.ts
- Add shared utilities for agent tools in agent-tools/utils.ts
- Simplify server.ts by moving error handling to dedicated module
- Update agent tools to use centralized error handling
- Improve type safety and error consistency across the codebase
