+++ packages/mcp-server/src/agent-tools/discover-dataset-fields.ts
@@ -1,6 +1,7 @@
 import { tool } from "ai";
 import { z } from "zod";
 import type { SentryApiService } from "../api-client";
+import { wrapAgentToolExecute } from "./utils";
 
 export type DatasetType = "events" | "errors" | "search_issues";
 
@@ -76,12 +77,12 @@ export function createDatasetFieldsTool(
           "Include example values for each field (set to false if you don't need examples)",
         ),
     }),
-    execute: async ({ includeExamples }) => {
+    execute: wrapAgentToolExecute(async ({ includeExamples }) => {
       return discoverDatasetFields(apiService, organizationSlug, dataset, {
         projectId,
         includeExamples,
       });
-    },
+    }),
   });
 }
