+++ packages/mcp-server/src/server.ts
@@ -35,107 +35,9 @@ import { logError } from "./logging";
 import { RESOURCES, isTemplateResource } from "./resources";
 import { PROMPT_DEFINITIONS } from "./promptDefinitions";
 import { PROMPT_HANDLERS } from "./prompts";
-import { ApiError } from "./api-client";
-import { UserInputError, ConfigurationError } from "./errors";
+import { formatErrorForUser } from "./internal/error-handling";
 import { LIB_VERSION } from "./version";
 import { MCP_SERVER_NAME } from "./constants";
-
-/**
- * Type guard to identify Sentry API errors.
- */
-function isApiError(error: unknown): error is ApiError {
-  return error instanceof ApiError;
-}
-
-/**
- * Type guard to identify user input validation errors.
- */
-function isUserInputError(error: unknown): error is UserInputError {
-  return error instanceof UserInputError;
-}
-
-/**
- * Type guard to identify configuration errors.
- */
-function isConfigurationError(error: unknown): error is ConfigurationError {
-  return error instanceof ConfigurationError;
-}
-
-/**
- * Formats errors for LLM consumption with appropriate telemetry handling.
- *
- * **Error Types:**
- * - User Input Errors: Clear guidance without telemetry
- * - Configuration Errors: Configuration guidance without telemetry
- * - API Errors: Enhanced messaging with HTTP status context
- * - System Errors: Full telemetry capture with event IDs
- *
- * @example User Input Error Response
- * ```markdown
- * **Input Error**
- *
- * It looks like there was a problem with the input you provided.
- * Organization slug is required. Please provide an organizationSlug parameter.
- * You may be able to resolve the issue by addressing the concern and trying again.
- * ```
- *
- * @example Configuration Error Response
- * ```markdown
- * **Configuration Error**
- *
- * There appears to be a configuration issue with your setup.
- * Unable to connect to sentry.invalid.com - Hostname not found. Verify the URL is correct.
- * Please check your environment configuration and try again.
- * ```
- */
-async function logAndFormatError(error: unknown) {
-  if (isUserInputError(error)) {
-    return [
-      "**Input Error**",
-      "It looks like there was a problem with the input you provided.",
-      error.message,
-      `You may be able to resolve the issue by addressing the concern and trying again.`,
-    ].join("\n\n");
-  }
-
-  if (isConfigurationError(error)) {
-    return [
-      "**Configuration Error**",
-      "There appears to be a configuration issue with your setup.",
-      error.message,
-      `Please check your environment configuration and try again.`,
-    ].join("\n\n");
-  }
-
-  if (isApiError(error)) {
-    // Log 500+ errors to Sentry for debugging
-    const eventId = error.status >= 500 ? logError(error) : undefined;
-
-    return [
-      "**Error**",
-      `There was an HTTP ${error.status} error with your request to the Sentry API.`,
-      `${error.message}`,
-      eventId ? `**Event ID**: ${eventId}` : "",
-      `You may be able to resolve the issue by addressing the concern and trying again.`,
-    ]
-      .filter(Boolean)
-      .join("\n\n");
-  }
-
-  const eventId = logError(error);
-
-  return [
-    "**Error**",
-    "It looks like there was a problem communicating with the Sentry API.",
-    "Please report the following to the user for the Sentry team:",
-    `**Event ID**: ${eventId}`,
-    process.env.NODE_ENV !== "production"
-      ? error instanceof Error
-        ? error.message
-        : String(error)
-      : "",
-  ].join("\n\n");
-}
 
 /**
  * Extracts MCP request parameters for OpenTelemetry attributes.
@@ -500,7 +402,7 @@ export async function configureServer({
                     content: [
                       {
                         type: "text" as const,
-                        text: await logAndFormatError(error),
+                        text: await formatErrorForUser(error),
                       },
                     ],
                     isError: true,
