+++ packages/mcp-server/src/tools/search-events/utils.ts
@@ -3,6 +3,7 @@ import { z } from "zod";
 import type { SentryApiService } from "../../api-client";
 import { logError } from "../../logging";
 import { UserInputError } from "../../errors";
+import { wrapAgentToolExecute } from "../../agent-tools/utils";
 
 // Type for flexible event data that can contain any fields
 export type FlexibleEventData = Record<string, unknown>;
@@ -117,41 +118,40 @@ export function createDatasetAttributesTool(
         .enum(["spans", "errors", "logs"])
         .describe("The dataset to query attributes for"),
     }),
-    execute: async ({ dataset }) => {
-      try {
-        const {
-          BASE_COMMON_FIELDS,
-          DATASET_FIELDS,
-          RECOMMENDED_FIELDS,
-          NUMERIC_FIELDS,
-        } = await import("./config");
+    execute: wrapAgentToolExecute(async ({ dataset }) => {
+      const {
+        BASE_COMMON_FIELDS,
+        DATASET_FIELDS,
+        RECOMMENDED_FIELDS,
+        NUMERIC_FIELDS,
+      } = await import("./config");
 
-        // Get custom attributes for this dataset
-        const { attributes: customAttributes, fieldTypes } =
-          await fetchCustomAttributes(
-            apiService,
-            organizationSlug,
-            dataset,
-            projectId,
-          );
+      // Get custom attributes for this dataset
+      const { attributes: customAttributes, fieldTypes } =
+        await fetchCustomAttributes(
+          apiService,
+          organizationSlug,
+          dataset,
+          projectId,
+        );
 
-        // Combine all available fields
-        const allFields = {
-          ...BASE_COMMON_FIELDS,
-          ...DATASET_FIELDS[dataset],
-          ...customAttributes,
-        };
+      // Combine all available fields
+      const allFields = {
+        ...BASE_COMMON_FIELDS,
+        ...DATASET_FIELDS[dataset],
+        ...customAttributes,
+      };
 
-        const recommendedFields = RECOMMENDED_FIELDS[dataset];
+      const recommendedFields = RECOMMENDED_FIELDS[dataset];
 
-        // Combine field types from both static config and dynamic API
-        const allFieldTypes = { ...fieldTypes };
-        const staticNumericFields = NUMERIC_FIELDS[dataset] || new Set();
-        for (const field of staticNumericFields) {
-          allFieldTypes[field] = "number";
-        }
+      // Combine field types from both static config and dynamic API
+      const allFieldTypes = { ...fieldTypes };
+      const staticNumericFields = NUMERIC_FIELDS[dataset] || new Set();
+      for (const field of staticNumericFields) {
+        allFieldTypes[field] = "number";
+      }
 
-        return `Dataset: ${dataset}
+      return `Dataset: ${dataset}
 
 Available Fields (${Object.keys(allFields).length} total):
 ${Object.entries(allFields)
@@ -173,12 +173,6 @@ ${Object.keys(allFieldTypes).length > 30 ? `\n... and ${Object.keys(allFieldType
 IMPORTANT: Only use numeric aggregate functions (avg, sum, min, max, percentiles) with numeric fields. Use count() or count_unique() for non-numeric fields.
 
 Use this information to construct appropriate queries for the ${dataset} dataset.`;
-      } catch (error) {
-        if (error instanceof UserInputError) {
-          return `Error: ${error.message}`;
-        }
-        throw error;
-      }
-    },
+    }),
   });
 }
